teach PCRE's '.' to ignore '\n' like perl
When '.' is used in a regexp, it matches all characters but the
ones defined as line delimiter, which is a configurable set in
PCRE, and with 2 of those sets not including '\n'.

perl allows for a configurable line delimiter string (not a set),
and therefore treats '\n' specially, preventing it for matching
regardless of what the delimiter contains, therefore when PCRE
uses one of those sets without '\n', the matches will differ:

 $ printf 'a\nb' | perl -n0le '/a.b/ or exit 1'; echo $?
 1
 $ printf 'a\nb' | pcre2grep -q -NNUL 'a.b'; echo $?
 0

Since the current behaviour for '.' is historical, a new compile
option has been invented to allow PCRE to match perl's behaviour
as an alternative.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>