pcre2_string_utils: avoid segfault with strlen(NULL)
pcre2(?:_dfa)?_match should return PCRE2_ERROR_NULL if the subject
is NULL, but the order that is done is incorrect, leading to crashes.

Workaround the issue by allowing strlen(NULL) to return a value
of 0, so it wouldn't segfault while trying to access a NULL subject
and therefore allowing the current check to be reached even if it is
done after the length of the subject is evaluated because it was
provided as PCRE2_ZERO_TERMINATED.

As a side effect, this also prevents crashes in pcre2_substitute
when the subject or the replacement string were NULL and the length
was provided as PCRE2_ZERO_TERMINATED and that would come out handy
if we want to be able to allow a NULL subject as valid when a length
of 0 was also provided.