pcre2_compile: make errorcode parameter optional
When the pcre2_compile interface was introduced, code was added to check
that BOTH the errorcode and erroroffset were not NULL, but usually only
one or the other is needed, in adition to a NULL return value to detect
failures.

Make the minimum change possible to allow passing a NULL and simplify
some of the calls used in pcre2_jit_test that will benefit from this
change.

A more complex change that would also avoid the multiple places where
checks for it were added has been punted from this change.