jit: fix pcre2_jit_free_unused_memory() if sljit not using allocator
sljit allows building without an internal allocator, but instead using
an external one.

make sure to only invoke the corresponding sljit call if an internal
allocator is in use (the default and as coded in pcre integration) to
avoid problems if the code is changed to use an external allocator
instead.