add pcre2_match_data_reset()
When reusing a match_data, there was always the possibility ot a
leak if PCRE2_COPY_MATCHED_SUBJECT was set, but since 10.41 the
heapframes are also stored in the match_data (explicitly so they
can be reused), but there is no way to clean them out unless
with `pcre2_match_data_free()`.

Add a new API that could be used to clear all additional data
from a `match_data` and that could be used independently to
improve the usability fot that reusing use case.