add pcre2_get_match_data_heapframes_size()
Since PCRE2 10.41, the match data contains a pointer to a vector of
frames allocated in the heap and that are used by pcre2_match()
when doing non JIT matches.

There is though, no outside visibility on the size of it, and therefore
the memory it uses is locked away until match_data itself is freed.

Add an API that allows getting that value, so an application could
decide based on its own experienced memory pressure to keep reusing
that match_data or not.

While at it, update the documentation of other related functions for
clarity.