pcre2test: use microseconds instead of milliseconds for timing
When using -t or its variants the timing results with a modern CPU
where not very significant even with 4 decimal digits.

While at it, move the factor calculation that can be done with
integer arithmetic and even optimized out for the more common case
when CLOCKS_PER_SEC == 1000000 out of the denominator.