add a C23 inspired checked integer multiplication helper
Compilers had become far more creative about optimizing code that
might be considered undefined behaviour, so improve our integer
overflow checking to prevent any possible miscompilations.

While at it, expand the use of the checks to pcre2test to fix a
similar problem to the one that was tackled in pcre2_compile.

Prefer (if available) a builtin provided by the compiler.