pcre2grep: avoid UCP mode for \d for compatibility with GNU grep
Matching multi-byte digits was released with GNU grep 3.9 and considered a
regression that required a couple more releases.

While implementations differ on their interpretation of `\d` under Unicode,
there is a recommendation[1] for a POSIX compatible mode that and a suggestion
that it will be better (safer and faster) to only consider ASCII digits for
that class (and [:digit]).

Use the recently introduced PCRE2_EXTRA_ASCII_BSD if `--posix-digit` is used
together with `-u` or `-U` for compatibility with GNU grep.

[1] https://unicode.org/reports/tr18/