windows: integrate visibility attributes without conflicts
When using a DLL in Windows, the function declarations (and definitions)
that are public are decorated with attributes but those declarations would
conflict with the ones that are selected when the compiler supports the
visibility feature.

Define instead a new macro that would be added to the corresponding
macros independently and while at it allow setting visibility with cmake.