avoid the possibility of dangling pointers to capitem in pcre2_compile
As shown by the following warning from gcc13:

  src/pcre2_compile.c: In function 'compile_regex':
  src/pcre2_compile.c:8319:17: warning: storing the address of local variable 'capitem' in '*cb.open_caps' [-Wdangling-pointer=]
   8319 |   cb->open_caps = &capitem;
        |   ~~~~~~~~~~~~~~^~~~~~~~~~

While normally the stack of "capitems" will be unwinded and the head of
it that is stored in "cb" set to NULL, this is not done in case of a
compilation failure.

Instead of keeping it inside "cb", initialize it in the stack of
compile_regex() and make sure it is gone once that function returns.