Fix and improve zig build
MAX_VARLOOKBEHIND now has to be set, so we do, which fixes builds on
latest version of pcre2. The build.zig was created for some older
version where this was not necessary. There are many many other
variables that we should set, but this is something. Baby steps.

The flag handling has been revamped, mostly to prepare it for having
varying flags set depending on input options and similar.

On Linux, Zig defaults to targetting the Musl C library. PCRE2, at least
pcre2grep only works with GNU libc and so we now rewrite the default to
make it work out of the box.

pcre2grep is now built when the -Dbuild_pcre2grep flag is specified.