Add a TH (title) macro to pcre2demo.3
per groff_man(7):

> A man page should contain exactly one .TH call at or near the
> beginning of the file, prior to any other macro calls.

This manual page didn't (and as a result `man --warnings` emits a
warning about the first `.SH` macro).

Make PrepareRelease add such a header to pcre2demo.3 along with a
comment noting that pcre2demo.3 is automatically generated (to avoid
any further confusion on that front).

This necessitates teaching 132html to know about that warning (which
we don't want to have appearing in the HTML output).

Also, update the EX and EE macros to reflect the current content of
`/usr/share/groff/1.22.4/tmac/an-ext.tmac`

Signed-off-by: Matthew Vernon <matthew@debian.org>