test: avoid UB in test17
Since 4ef0c51 (Interpret NULL pointer, zero length as an empty
string for subjects and replacements., 2021-11-30), a NULL pointer
with zero length has been internally replaced with an empty string
in all matching functions, but the test for jit was incorrectly
using the fastpath which intentionally skips all validation and
results in NULL pointer arithmetic operations.

Change the test to use the validating jit path instead.