maketables: avoid misleading values in case flipping table
The tables generated by pcre2_maketables() include one that maps
all lowercase characters on the first 255 code points to their
corresponding upper case code point, but fails to notice that
toupper() could return a larger code point and therefore result
in the store of a truncated and unrelated code instead.

Restrict all values to what is valid for uint8_t and document
in the test case the failure for character 'μ'[1] (U+00B5) and
that was incorrectly getting back 924 (U+039C) from macOS fr_FR,
and resulting in an incorrect case equivalent with the truncated
value of 159.

[1] https://en.wikipedia.org/wiki/Mu_(letter)