pcre2posix: make regex_t.re_match_data slightly less opaque
Document its existance and add a re_match_data_foreign field that
could be set to instruct the code to not create and destroy one
as needed.

The logic uses a "private" re_match_data_owner field to make it a
little more resilient against the possibility of them being set
accidentally by an uninitialized regex_t, but obviously documents
that only the use of initialized (to 0 at least) regex_t is safe.

On the same line, and while not strictly needed, the match_data
pointer is set to NULL after releasing it when the flag is not
enabled.