pcre2posix: avoid snprintf quirks in regerror
C99 snprintf could return a negative value on error, so detect
that and correct the return value and message as needed.

There is also an older variant that returns characters used, and a
Microsoft one that does something similar but without the NUL
termination, so account for those as well.

While at it, remove the padding on the message and add a helper
function to give an accurate estimate of its length when printed.

Note that the current logic ensure that only possitive ints will
be provided to the helper function.