m4/pcre2_visibility.m4: add basic support for non GCC compilers
Assume that a compatible `-fvisibility=hidden` option is available
and that the compiler handles correctly `#warning` macros.

Restrict the visibility test if `-Werror` is unrealiable to protect
for compilers like xlc or systems like cygwin that might allow the
compiler flag without errors but wouldn't work.

While at it, lower the minimum autoconf required to the one that
introduced the latest macros and tested in OpenBSD 7.5 with automake
1.9.