pcre2_compile: belt and suspenders for check_lookbehinds()
Maybe pacify a coverity report by making sure a safe erroroffset
is reported for an error case that should never happen.

To keep the usefullness of the crash that could be triggered in
case of a bug, reintroduce the original value for DEBUG builds.

While at it, cleanup code from a related function that was not
needed and change a cast to better match the current size as
well as a relevant typo.