doc: include summary of directives in pcre2_set_optimize.3
Make the documentation of the new API more useful at a first glance
by providing the list of values that can be used while keeping all
details in pcre2api.3, and just like it is done in other similar pages.

While at it reorder the entries for the directives in pcre2api.3 so it
is more natural and to match the one used here.