Update soversion passed to libtool for 10.45 (#636)
I believe this is the correct procedure, based on Philip's documentation.

The libpcre2-posix interface is completely unchanged, but the source code has been updated.

The libpcre2-NN interface has been extended in backwards-compatible ways, with new enum values and API functions, so callers compiled and linked against the old version may use the newer as a drop-in replacement, but callers compiled against the 10.45 headers will fail when used against the old version (if they require any newly-added functions).