Fix UCP-only handling of caseless character classes (#644)
Fixes #526

The handling is rather brittle; but it will do for now. We should always compute the caseless extension of character classes, for cranges and bitset branches; however we can add some delicate and dubious optimisation to avoid doing the caseless extension two times, when converting from a cranges to a bitset.