Try new  strategy for auto updates (#664)
I previously added a CI that the documentation was not dirty. This turns out to be too strict: it's onerous to have contributors install the exactly-matching `nroff` setup required to sync the documentation.

Instead, the CI check now commits and pushes the change to the PR (if it's a same-fork PR), and ignores the changes if it's a PR from another fork.

The CI will push the PrepareRelease changes on master, if needed.

Fixes #653