Add tests using nm to check so contents (#693)
We want to ensure that the symbols included in shared object files (.so, .dylib, .dll) are as expected.

This verifies that our Autoconf and CMake builds are using the `-fvisibility` flags and attributes correctly.

As it turns out, the CMake build on Solaris is not working, and was exposing internal symbols due to a CMake issue.