Fix a build issue due to previous CMake change
The problem is that the Readline/zlib include paths were being included
before our own include path (wrong import order). This caused a problem
if the zlib include path was /usr/local/include, and that directory
happened to unfortunately contain a pcre2.h file for an (older) version
of PCRE2.