Add a simple .editorconfig file (#747)
This ensures that (most) compatible editors will insert tabs into the makefiles and spaces elsewhere. We also don't want trailing whitespace to be stripped when we edit the test files. Finally, it helps to open testinput1 in Latin-1 and testinput2 in UTF-8 mode. All this can be automated.

The editorconfig file should work with any text editor - although note that VS Code needs an extension to make use of it.