Add test output to make erroroffsets more readable (#750)
Currently, pcre2test prints the erroroffset as a simple number, which makes it hard to see in the test output where it's actually pointing.

I have added printing of the relevant portion of the pattern, to show visually where the offset points.

There is also a new function error_direction() where I have started to classify which PCRE2 errors point to the left or right of the erroroffset. When I have fixed this all to be consistent to my satisfaction, I will document this so that clients have a better way to indicate which character is wrong in a regex.