pcre2_config: add PCRE2_CONFIG_EFFECTIVE_LINKSIZE
PCRE2_CONFIG_LINKSIZE can be used to determne the linksize that the
library was configured to use, but the actual size (and therefore
the maximum size of patterns that can be processed) is not visible.

Add PCRE2_CONFIG_EFFECTIVE_LINKSIZE to gather that value and use it
instead of PCRE2_CONFIG_LINKSIZE in `pcre2test` for reporting
`linksize`; mark the old setting as "deprecated" and move the logic
to derive one from the other out of the API documentation.