maint: introduce LintMan to aid on tracking/updating values
Allow tagging the documentation with a `#define` value that could
be then updated programmatically.

Update the value for MAX_NAME_SIZE in pcre2limits.3 that was missing
since ced3b0f0 (Increase name length to 128, 2024-03-11) and while
at it, improve on its description and a tag for a related variable.

For completeness, add also a tag to the same value in pcre2pattern.3
and the configuration for VMS that was missing since 6c670c78
(Update overlooked cmake update of name size to 128, 2024-03-11)