pcre2_match: improve `match_ref()` logic for robustness
A recently fixed bug in Scan Substring triggered problems in the
current logic that would result in crashes and the skipping of
safewards.

Add an assert to validate that this function is never called with
an `eptr` over the end of the subject and use the correct return
value for non partial matches of back references.