build.zig: update to zig 0.15 (#777)
Also fixes dynamic builds (even in Windows) and improved
compatibility by enabling compiler builtins and visibility.

While at it, change the JIT flag to `support_jit` for compatibility
with cmake, update CI to build with JIT and use the `bigstack` mode
that is required with debug builds using clang with sanitizers.