pcre2_config: address inconsistency with the linksize reported in non 8-bit libraries (#775)
pcre2_config: add PCRE2_CONFIG_EFFECTIVE_LINKSIZE

PCRE2_CONFIG_LINKSIZE can be used to determine the linksize that the
library was configured to use, but the actual size (and therefore
the maximum size of patterns that can be processed) is not visible.

Add PCRE2_CONFIG_EFFECTIVE_LINKSIZE to gather that value and use it
in `pcre2test` for reporting `linksize`; move the logic to derive
one from the other out of the API documentation and test scripts to
improve maintainability and while at it enable test8 in Windows for
all link sizes.