De-macro the pcre2test code (#797)
The current style in pcre2test.c is rather hard to work with. Every little call to PCRE2 functions has to be guarded behind many layers of complicated macros like "SET", "CASTVAR", and more.

At the cost of a modest increase in binary size, I would like to split out roughly half of the logic in pcre2test.c into a separate file, which can be multiply-included.