Backport of pcre2-10.47-Fix-for-callback.patch
Cherry-pick of cc6c84d0207d9d09ca9e42996f0bc50353c03157

Add fix and tests for callback enumerate (#801)

This is a bug in PCRE2 10.45 and 10.46, not present in 10.44 and earlier.

The bug is that applications using pcre2_callout_enumerate() will crash badly on patterns with Unicode (ie non-ASCII) character classes. There is an out-of-bounds read of arbitrary length, including misinterpreting the bytes as offsets, allowing the read to jump forwards in memory to pretty much anywhere.

Applications which call pcre2_callout_enumerate should potentially downgrade to PCRE2 10.44 until we release an update of PCRE2 with the fix.

Since this function is (clearly!!) not used often by applications, I am not currently treating this as very high severity. There is no way for an attacker to make any application call this function, if it is not currently using it.

The root cause seems to be commit 24f9d8d (#540).

---------

Co-authored-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>