jit: enable x86 SIMD helper if SSE2 is available
The current implementation needs SSE2 support (which is part of the
base amd64 specification and available in most CPUs from this century.

Without this change, SIMD will be only enabled if SSE4.1 is detected
which is not true for older AMD cpus, even if 64bit capable.

While at it, update related documentation and other minor tweaks.