Add runtime checks for invalid uses of \K in lookaround (#812)
If the match start is before the start-offset, or if the match ends before it begins, then pcre2_match() and JIT match reject the match with a new error (unless the option is active to allow \K within lookaround).

The only way to get in this situation should be when someone uses a subroutine call or similar control flow to jump to a \K from inside a lookaround.