fixes for 10.47 (#819)
* update required autoconf version

Since abc2ae68 (Update AX_PTHREAD (#694), 2025-02-12), when using
autoconf 2.60 with autogen, shows:

  error: possibly undefined macro: AS_ECHO

Use 2.62 instead, which include that macro, as the new required version.

* jit: enable x86 SIMD helper if SSE2 is available

The current implementation needs SSE2 support (which is part of the
base amd64 specification and available in most CPUs from this century).

Without this change, SIMD will be only enabled if SSE4.1 is detected
which is not true for older AMD cpus, even if 64bit capable.

While at it, update related documentation and other minor tweaks.