pcre2test: keep track of internal API with sljit platform names
SLJIT platform names used to be 64 bytes wide, but had since
grown.

Add a new header to keep track of those internal API numbers
and use that to size the needed buffers accurately.

While at it, do the same for the standalone JIT test which
used by default a more realistic power of 2 buffer size.